backend/requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==4.2.17
django-cors-headers==4.7.0
django-filter==25.1
django-unfold==0.54.0
djangorestframework==3.16.0
filelock==3.16.1
mysqlclient==2.2.7
pillow==10.4.0
platformdirs==4.3.6
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.0.1
rsa==4.9
sqlparse==0.5.3
Telethon==1.39.0
typing_extensions==4.12.2
virtualenv==20.29.1
